--- Projects/TEAM-2/Tsihanchuk/Task_201/201.cs
@@ -1,17 +1,0 @@
-using System;
-
-namespace AVG
-{
-    class Program
-{
-    static void Main(string[] args)
-    {
-        string first = Console.ReadLine(); 
-            string second = Console.ReadLine();
-
-            float avg = (Int32.Parse(first) + Int32.Parse(second)) /2;
-
-           Console.WriteLine(avg);
-    }
-}
-}
