+++ Projects/TEAM-2/Tsihanchuk/Task_201/Task_201.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32328.378
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task_201", "Task_201.csproj", "{E709E82A-7540-43C2-9B93-A6EBA48287F9}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E709E82A-7540-43C2-9B93-A6EBA48287F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E709E82A-7540-43C2-9B93-A6EBA48287F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E709E82A-7540-43C2-9B93-A6EBA48287F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E709E82A-7540-43C2-9B93-A6EBA48287F9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {42498549-ACE2-46AE-BD7D-8A3CB438C97F}
+	EndGlobalSection
+EndGlobal
